refactor(mvvm): extract ObservableValidator from ViewModelBase
- Move INotifyDataErrorInfo + DataAnnotations validation infrastructure
  (HasErrors, GetErrors, ValidateProperty, ValidateAllProperties,
  AddError/ClearErrors, OnErrorsChanged, the validation cache and the
  PropertyChanged hook) out of ViewModelBase and into a new
  ObservableValidator base class derived from ObservableObject
- Re-base ViewModelBase on ObservableValidator so existing consumers
  keep their full validation API; the UI-state members (IsEnabled,
  IsVisible, IsBusy, IsDirty, IsSelected) and MessengerInstance stay on
  ViewModelBase
- Override ShouldSkipValidationOnPropertyChanged on ViewModelBase to
  exclude the UI-state properties from validation re-runs (preserving
  prior behavior; previously the skip list was inlined in
  OnPropertyChangedValidation)
- Update ObservableDtoViewModelAttribute XML doc cref to point at the
  new ObservableValidator.HasErrors location
- Add ObservableValidatorTests covering HasErrors lifecycle,
  ValidateAllProperties on init, auto-validation on property change,
  ErrorsChanged dispatch, HasErrors PropertyChanged raises, and a
  smoke test that the class works without ViewModelBase
